Extract to its own file.
